tests/ProtoR.Domain.UnitTests/SchemaGroupAggregateTests/RulesTests/ProtoBufRulesTests/EnumAddedRuleTests.cs
namespace ProtoR.Domain.UnitTests.SchemaGroupAggregateTests.RulesTests.ProtoBufRulesTests
{
    using System.Diagnostics;
    using ProtoR.Domain.SchemaGroupAggregate;
    using ProtoR.Domain.SchemaGroupAggregate.Rules.ProtoBufRules;
    using ProtoR.Domain.SchemaGroupAggregate.Schemas;
    using ProtoR.Domain.UnitTests.SchemaFixtures;
    using Xunit;

    public class EnumAddedRuleTests
    {
        private EnumAddedRule rule;

        public EnumAddedRuleTests()
        {
            this.rule = new EnumAddedRule();
        }

        [Theory]
        [RuleTestData("EnumAdded", "AddedEnum")]
        public void Validate_WithAddedEnum_ShouldNotPass(ProtoBufSchema a, ProtoBufSchema b)
        {
            ValidationResult result = this.rule.Validate(a, b);

            Assert.False(result.Passed);
        }

        [Theory]
        [RuleTestData("EnumAdded", "InnerEnum")]
        public void Validate_WithAddedNestedEnum_ShouldNotPass(ProtoBufSchema a, ProtoBufSchema b)
        {
            ValidationResult result = this.rule.Validate(a, b);

            Assert.False(result.Passed);
        }

        [Theory]
        [RuleTestData("EnumAdded", "NoAddedEnum")]
        public void Validate_WithNoAddedEnum_ShouldPass(ProtoBufSchema a, ProtoBufSchema b)
        {
            ValidationResult result = this.rule.Validate(a, b);

            Assert.True(result.Passed);
        }
    }
}
